delete blank without exception

+++ src/main/kotlin/at/birnbaua/sudoku_service/jpa/jpaservice/JpaService.kt
@@ -63,7 +63,9 @@ abstract class JpaService<T: Any,ID: Any> (private val repository: JpaRepository
     }
 
     fun deleteById(id: ID) {
-        repository.deleteById(id)
+        if(repository.existsById(id)) {
+            repository.deleteById(id)
+        }
     }
 
     fun delete(entity: T) {
